tests: Fully implement gobject-introspection test suites
This implements all the possible tests from the gobject-introspection
test suites GIMarshallingTests, Regress, and WarnLib. For anything that
is not supported or buggy in GJS, we still write a test but skip it,
adding the bug link as the skip message.

The point of this is that we can get as much test coverage of the GJS
codebase as possible by using the tests that already exist for
gobject-introspection. Doing this has already yielded a number of bug
reports which would be suitable for contributors new to GJS to work on.

After merging this we'll also be able to use our code coverage report to
get a clearer picture of what parts of the GJS codebase aren't exercised
at all, so where we should add more testing.