heapgraph: Add ability to label nodes
The profusion of Object or GObject_Object nodes can be quite confusing.
If you add a __heapgraph_name property to an object in your code with a
string value, then it should be visible in the heap dump, and heapgraph
will now detect this situation and display it as a label.

    myObj.__heapgraph_name = 'My object';