heapgraph: Always keep "interesting" labels in graph
If you specify an edge target, then it may happen that the edge isn't
actually printed in the graph. Since the search only prints at most one
reason why a node is reachable, then for example if the edge's source
and destination are rooted then the edge won't appear.

In the console output tree graph we can't handle multiple edges pointing
to a node, but a dot graph can handle that just fine. So we add these
extra "interesting" edges at the end of the code that prints out the dot
graph.