heapgraph: Add an argument to print out unreachable nodes
In any given heap dump there seem to be a bunch of nodes that are
printed out as a result of visiting arenas, that are not reachable from
any root. For these nodes the breadth-first search returns no path.

It's a bit confusing to see, for example, that 25 nodes of type
GObject_Object were found in the heap dump but only 8 are printed out.
The others are unreachable. This adds a --show-unreachable (-u) switch
that prints out the unreachable ones at the end of the tree. In the case
of dot output, it prints them with a dotted border.

I think mostly this switch is not going to be very useful, but it
happens to be useful for me right now because I'm augmenting the graphs
with GObject reference information from C.