context: Collect nursery before heap dump
Until now, mostly we didn't care about the nursery because only plain JS
objects are created there; objects with finalizers, which comprise all of
our custom objects in GJS, are automatically promoted to the tenured heap
in SpiderMonkey's implementation of generational garbage collection.

However, with the taking-out-the-garbage patches we will create a lot more
Proxies (GObject wrappers) and plain objects (GObject expando objects)
which can be allocated in the nursery. This can result in them being
missing from heap dumps when we use the IgnoreNurseryObjects flag to
js::DumpHeap().

The alternative is the CollectNurseryBeforeDump flag, which promotes any
live objects in the nursery to the tenured heap. This way, we are
certain to capture all of our wrappers and expando objects in the heap
dump. The downside is that dumping the heap can now reduce performance.
But it is something that should only be done for debugging, anyway.

So here's this commit in advance of the taking-out-the-garbage branch, so
that I don't have to keep rebasing it.

See GNOME/gjs#217.