CI: Organize build jobs into recommended, minimal, and maximal
Since jobs fail a lot on our CI, it makes more sense to build fewer of
them. This renames the "fedora" job in the first phase to
"build_recommended" and makes its arguments explicit. In the second
phase it consolidates all of the builds with feature X turned on or off
into two "build_minimal" and "build_maximal" jobs.

It turns out that we have not been building the profiler at all in CI
recently because the sysprof-capture-3 package isn't available. Disable
it explicitly for now, with a fixme note.

For this reason we also have to run build_maximal on Fedora, since there
isn't a sysprof package available for Ubuntu yet. We run build_minimal
on Ubuntu instead.