build: Remove xvfb-tests option
This option is mainly used on headless systems such as CI. It makes more
sense for those wanting to run the tests in a headless configuration, to
supply their own xvfb server with the desired options, instead of
relying on GJS to get the options right.

This commit removes the option from configure.ac and Makefile-test.am,
and amends the CI scripts to run "make check" under xvfb-run.