object: Throw if constructing an unregistered object inheriting from GObject
When a class inherits from a GObject class, it must be always registered
using GObject.registerClass, however gjs was accepting this silently,
until the object wasn't passed around.

To make this clearer, verify if the object's new.target is set and
if it has the '$gtype' own property defined. Throws otherwise.

Signed-off-by: Marco Trevisan (Treviño) <mail@3v1n0.net>