CI: Add meson CI job
During the transitional period when we have both Autotools and Meson
build systems, we should have a Meson job in order to make sure that
a change to one build system doesn't break the other.

When removing the Autotools build system, this job should be removed and
all the other jobs should be ported to use Meson.