private: Don't pass argv to for_js_typecheck and use result for later
When using GObject.registerClass to register a class that doesn't inherit
a GObject type gjs was throwing a too generic error that wasn't much clear.

I.e.:
 JS ERROR: TypeError: GObject_Object.prototype.register_type called on
                      incompatible Object

Since we don't care about the callee here but more the object types,
use the simpler version of for_js_typecheck() that relies on
gjs_typecheck_instance() to get a typed error message.