GObject: Optionally use the caller script basename to define the GType
When computing a GType we're only using the class name, but this
might lead to clashes when the same name is used various times
in a project.

To reduce this possibility, add a GObject property
(GObject.gtypeNameBasedOnJSPath) that can be set to true to make gjs
try to compute the caller script path, by getting the first call
coming from another script in the error stack (unfortunately it seems
the only way in JS :().

Then use this value to generate the GType name.