build: Bump gobject-introspection requirement to 1.61.2
Porting to Meson exposed a bug in the gobject-introspection test file
warnlib.c, which is installed but includes config.h. This bug was fixed
in gobject-introspection 1.61.2.

The bug was hidden by the autotools build, which just tacks the same
include directories onto everything that's compiled, so it erroneously
picked up GJS's config.h.