CI: Remove non-native architecture builds
In practice, these were not being run very often, and when they did run
it was often mystifying to debug if they failed. Nonetheless, Docker
images were built every month for them, which took 5 hours each, rather
than half an hour as for the rest of the images. Additionally, often the
Docker images failed to build in the first place, as Fedora's GCC on that
architecture would often turn up a bug while trying to compile
SpiderMonkey.

In general I have come to believe that it's downstreams' responsibility
to ensure things work on any weird architectures that they ship on.
However, it seems that we do have a native aarch64 runner in GNOME
GitLab, so we can at least make use of that in the future, using less
electricity than the current QEMU setup.