tests: Fix one remaining cppcheck error
sscanf() is basically unsafe to use, but here we are parsing well-formed
LCov data in the test suite, so there's no user input involved. We can
assume that source files in the test data have 99999 lines or fewer, and
we dynamically generate the field width for the function name to fit the
buffer we have allocated.