CI: Port test-ci.sh to POSIX shell
In the next phase we are planning to build the linter Docker image on
Alpine Linux instead of Fedora, which includes busybox by default
instead of bash. In order to keep the Docker image light, we don't want
to install bash on it; use only POSIX shell syntax in the test-ci.sh
script.