GLib: Override string manipulation functions
There are a number of string manipulation functions in GLib that return
a pointer to the same passed-in string. These can't be annotated
properly, and will mostly crash. We want to prevent user code from
calling them, and so we override them.

In the overrides we provide approximate implementations of each function
so that if they had happened to work in the past, they will continue
working, but log a stack trace and a suggestion of what to use instead.

Closes: #283