CI: Skip linter jobs if no linted files modified
This uses GitLab's changes expressions to skip the linter jobs if no
files were modified in the code being pushed that would be linted by
that job.

So we only run cpplint and cppcheck for changes to .c, .h, and .cpp
files, and we only run eslint for changes to .js files.