jsapi-util-root: Add JSContext parameter to switch_to_unrooted
switch_to_rooted required already a JSContext parameter to apply,
how to unroot we used the current context by default.

However this implies that a generic api as GjsMaybeOwned will
depend on GjsContextPrivate that is instead unrelated to this.

So, just take a context parameter, and reflect this in the calls