jsapi-dynamic-class: Switch to RAII classes
We are going to refactor this anyway when we remove JSRequest in
SpiderMonkey 68. So, use RAII classes here and get rid of the gotos. The
comment at the top of the function about using 'volatile' to force the
variables to be collected by the GC has been inaccurate since
SpiderMonkey switched to an exact garbage collector several years ago.