meson.build: Don't check for GCC-specific items on MSVC
Instead, we use the msvc_recommended_pragmas.h as we do in the GTK
stack, as well as in conjunction to disabling some other warnings,
as we do in the the NMake Makefiles, so that we cover more or less the
same items in MSVC builds.

Since there is no -Bsymbolic-functions equivilant in MSVC, we just skip
checking and using that as well.

Also, make sure that we pass in the additional flags that we need for
Visual Studio builds, as we did in the NMake Makefiles.