gjs/*.cpp, util/*.cpp: Check for sys/signal.h and _WIN32
Add a configure-time check for sys/signal.h and only include it in the
sources when it is found.  For context.h, since it is public, only
include it when we are not building for Windows (i.e. _WIN32 is not
defined), because it does not exist for Windows.

Also, we might be checking for G_OS_WIN32 before we include the GLib
headers, so check for the presence of _WIN32 in those cases, since it
will be defined by default on all compilers targeting Windows.