arg: Support flat arrays of structs
This is actually possible without a cleanup list, contrary to what was
previously asserted in #44.

It's possible to pass a flat array of C data if the array element type
is a struct or union not registered as a boxed type. In this case,
g_type_info_get_pointer() will be false for the GI info of the function
argument. We detect this case and create a flat array where the structs
are copied by value.

We preserve the existing special case for flat GValue arrays, since we
need to use gjs_value_to_g_value() there instead of
gjs_value_to_g_argument(), and doing both in the same function would
make the code less readable.

See #44.