build: Handle case where $bindir/gjs symlink already exists
Python's os.symlink() will fail if the symlink already exists, so when
installing GJS in a prefix where it was already installed. This was
previously avoided with the -f option to ln.

Instead, create a temporary link and atomically rename it over the
existing link.

Unreviewed, pushing to fix build.