tests: Test for empty flat arrays of structs
We should be able to pass null or [] to a parameter expecting a flat
array of structs. This case was inadvertently broken on the stable
branch when we prevented the crash as part of #44. Even though on the
development branch we now support these kinds of arrays, we should still
test passing empty ones.

Closes: #289