js: Remove most goto error handling
This is C++, we have RAII, so while "goto out" is a good pattern in C if
you are not using g_autoptr, it's not a good pattern in our code. This
removes most goto error handling in favour of RAII or splitting
non-cleanup parts out to a second function. There are still a few
instances of it left, that will require more extensive refactoring.