build: Don't error out on warnings in test GIR sources
Even if compiling with -Werror, we don't want to cause an error if the
test GIR sources (regress.c, gimarshallingtests.c, warnlib.c) cause a
compiler warning. These sources are provided by gobject-introspection
and so the versions of sources showing up in our CI system is dependent
on when gobject-introspection decides to make a release.