gtype: Don't use lookupForAdd() when adding GType to table
This was a bug exposed by more efficient garbage collection and better
debug assertions in SpiderMonkey 68. We cannot use lookupForAdd() when
checking if a GType is already present in the GType table (which we did
to anticipate adding it if it wasn't present) because creating the GType
wrapper object might trigger a GC, which could cause another GType to be
garbage collected, which would remove it from the GType table. Mutating
the table invalidates the AddPtr returned by lookupForAdd(), which is a
bug.