README.MSVC.md: Update build instructions
Let people know how SpiderMonkey 68 can be built, since it now requires CLang
for compilation, and additional ways to obtain the SpiderMonkey sources.

Unfortunately, some GCCisms and CLangisms creeped into the public headers
of the SpiderMonkey 68 public headers, so they must be updated first before
carrying out the build of GJS, which are being described here, and it is
currently not possible to build GJS with CLang on Windows.

Note that since CLang link to the same CRT DLLs as Visual Studio 2017 and 2019,
it is perfectly fine to build and link GJS against SpiderMonkey 68 with
the Visual Studio 2017/2019 builds of GJS, when the needed updates to the
SpiderMonkey headers are applied, as described above.