tests: Avoid filename conflict when tests run in parallel
It seems that it's possible for the profiler start/stop test to have run
during the script test that verifies that the profiler doesn't write a
file, which causes it to fail because they both use the default profiler
file name. Give the start/stop test a different file name and delete it
afterwards.