object: Fix dangling pointer crash
The linked list is used by update_heap_wrapper_weak_pointers() which is
responsible for clearing weak pointers to the ObjectInstance wrapper
that become dangling.

There's a chance the GC sweeps the wrapper between the call to
gobj_dispose_notify() and the ObjectInstance being destroyed. The GC
will call the weak pointer callback when it does this.

If the ObjectInstance is missing from the list at this point (because of
the unlink()) then the weak pointer will be missed and become dangling.
Afterwards, when the ObjectInstance is being destroyed, the Heap
destructor will trigger the write pre-barrier, causing a crash.

Closes: #301