object: Separate closures and vfuncs
Previously, signal connections, pending notify callbacks, and vfuncs
were stored in one list that was a member of ObjectBase. However, only
ObjectPrototypes have vfuncs, and only ObjectInstances have signal
connections and callbacks. This commit separates them, so that later we
can move the signal connections and callbacks to expando objects as part
of #217.