wrapperutils: Move is_custom_js_class()
This name of GIWrapperBase::is_custom_js_class() is misleading. If there
is no GI introspection info, then the class may be a custom JS class,
but it could also be an internal class not exposed through
introspection, like GLocalFile.

In most cases is_custom_js_class() can just be replaced with !info().

In one case of existing usage of is_custom_js_class(), we really do mean
it, in add_property_impl(). For this we check the custom_type_quark() on
the GType. For better readability, this adds an is_custom_js_class()
method to ObjectBase which actually does what it says.

See also #299 and commit 74171bf3ae91ff64337b6261fe024e34e073f86f.