object: Return early from property resolve if name starts with -
This is a small optimization. GObject property names must start with a
letter. Any other character gets turned into a dash by
canonicalize_name(), so if the property name starts with a dash after
canonicalization then it's not a valid property name and can't be a
GObject property, so we can return early from any code that is checking
whether it is or not.

I noticed that resolve was checking for _init and _instance_init, so
this seems like a pretty common case to optimize for.