object: Only define interface overridden properties on the owner type
The previous code did not check if overridden interface properties were
actually overridden by the type whose ObjectPrototype it was defining
them on. So, a subclass of a type that implemented an interface would
get the interface's properties defined on its prototype. In most cases
that didn't matter, but in the case where the implementing type was
defined in JS, then the subclass wouldn't get the implementing type's
property accessors.

Closes: #306