arg: Fix GList and GHashTable marshalling on 64-bit big-endian
It is not correct to simply write to arg->v_pointer and read out a
different member of the GIArgument union, or vice-versa. We were doing
this when stuffing integers into pointers for storage in a GList,
GSList, or GHashTable. In the case of 32-bit ints, this would read out
the wrong half of the pointer on big-endian 64-bit architectures.

Closes: #309