repo: Remove JSClass macros
The macros in jsapi-class.h are not needed here. Instead of using
JS_InitClass() which requires a constructor and prototype object, we can
just use JS_NewObject() which will use a plain object as the prototype,
and we don't need to store private data in the instances since there is
none.