function: Ensure either JS or GArgument return value, not both
In gjs_invoke_c_function it should be the case that either the JS return
value is wanted, or the GIArgument return value, not both. Ensure this
with an assertion, and use only one condition throughout the body of the
function.

Use type inference to clean up the invocations of gjs_invoke_c_function
a bit.