function: Merge two ways of invoking constructor from C
We had two separate ways of invoking a JS constructor function, one was
used in union_new and the other in Fundamental::invoke_constructor. The
original one, gjs_invoke_c_function_uncached(), was added by Colin in
commit 47381c920222f0502a9328b664669edf9fc8c6ca and was intended for
constructors that didn't need the fast path of caching their function
data.

The other, gjs_invoke_constructor_from_c(), was added later for
fundamental type support and did cache the function data, but did not do
argument marshalling on the return value, just returning the C pointer
instead.

Both of these features are useful for both situations, so we merge the
two functions into one gjs_invoke_constructor_from_c() which does not
cache the function data and does not do argument marshalling on the
return value.

Since we now invoke the constructor by its GIFunctionInfo, we don't need
to store its name on FundamentalPrototype, which allows us to remove
actually a lot of code for storing and tracing the property key.