function: Use JS::CallArgs throughout invocation call chain
If you have a JSObject for the 'this' object, a JS::Value array for the
arguments, and a mutable JS::Value for the return value, then that's
basically a JS::CallArgs.

The only difference is that when calling a constructor, you may not be
able to get the 'this' object from args.computeThis(). For
gjs_invoke_constructor_from_c(), we need to provide the 'this' object
that was provided to the constructor.