function: Improve format_function_name()
It's not necessary to pass in the information about whether the function
is a method or not, it's available from the function's GICallableInfo.

Fix the type of function->info to be GICallableInfo. In some cases it's
not a GIFunctionInfo, as can be found out by trying to call a
GIFunctionInfo method on it unconditionally.