cairo: Use cairo_t as the private pointer for Cairo.Context
Instead of allocating a separate private struct that allocates a cairo_t,
use the cairo_t directly as the private struct. This makes the code
simpler and saves 8 bytes per object. The only thing to watch out for is
that the pointer can be null if it was already disposed.