arg: Automatically figure out the proper container type for a JS value
Add template functions to get the JS value to a variable that is big
enough to fit the requested type, doing also checks in case the
resulting value is out of bounding for the requested type.

This allows to redefine hashtable_int_key() so that there's no type
checks and will allow to get values per type.

Move the functions to a different header not to bother arg with such
stuff that might be re-used also by other code paths