arg: Automatically set all the numeric values to GIArguments
Use a template-function to convert all the JS Values to the proper c++
type and ensure they're within the valid ranges.

As per this we need to use a more relaxed conversion to unsigned types using
doubles not to have implicit conversion of negative types.

Use a function override in order to keep a more generic version as
inline implementation, while we specialize the out-of-range error
handling per type