build: Postprocess IWYU output
When running IWYU, the output is really hard to read. Ideally we want
only a diff of which lines to add or remove for each file, and no output
for a file that needs no changes.

Also take this opportunity to implement the "fwd-decls" header
js/TypeDecls.h in the postprocessor so that we don't get requests to
forward-declare JSContext everywhere; and make the postprocessor
suppress false positives.