arg: Implement gjs_array_to_strv using the auto-array
Even in this case we can just implement this using a specialized
implementaiton of js_value_to_c.
However, since the array contains new strings we also need to templatize
the array allocation and free functions, so that we memzero it and we deeply
free it in case we handle a GStrv.

As per this, only set the last element to NULL on non-string arrays