jsapi-util: Use string vector in gjs_define_string_array
This makes the code cleaner and takes care of memory management. We
propagate this API change back to all callers as far as possible, as
well. (gjs_context_define_string_array() is a public C API, so its
signature can't be changed; and GjsContextPrivate::m_search_path must
be a POD type because the GObject property setter sets that memory
location before the GjsContextPrivate constructor runs.

Co-authored-by: Evan Welsh <noreply@evanwelsh.com>