tests: Adapt to GTK4 API changes
The monthly CI image build included a newer version of GTK4 which had
some API changes. The one we need to contend with is the removal of
gtk_container_add() in favour of widget-specific functions such as
gtk_window_set_child().

The destroy vfunc was also removed so it seems that one avenue to get
a JS callback called during garbage collection no longer exists in GTK4,
so we can remove that test.