CI: Only setup the builddir in run_iwyu
There's no need to build the whole project, as we are only analyzing the
source code. We just need to run 'meson setup' on the builddir so that
we have the compilation database.

Also print the list of affected files after running Meson so they don't
scroll out of sight.