arg-cache: Limit GI arguments to 253 and use 8-bit storage
It's wasteful to use a signed int to store the number of function
arguments. We do need two special values to indicate return value and
instance parameter, but these can be the two highest unsigned values
instead of -1 and -2.

It seems reasonable to use 8 bits for the number of arguments. Throw an
exception when invoking a function with too many arguments, and use
assertions in GjsArgumentCache for internal checks.