arg-cache: Save space in enum bounds
We were storing a 64-bit int for the min and max values of an enum type.
However, the documentation of gobject-introspection notes that enum
values are always representable in 32 bits, either signed or unsigned,
and the use of a 64-bit return type is to easily allow for both.

We store the bounds in 32-bit unsigned bitfields, and use a flag that we
add to the slop space in GjsArgumentCache (which we can share with
number types as well) to indicate whether the bounds should be
interpreted as signed or unsigned.