tests: Fix specifics in tests to match SpiderMonkey 78 output
The LCOV output has changed slightly again in SpiderMonkey 78, so we
must update our tests to match. There are trivial modifications such as
changing the order that function records appear in, but it also seems
that SpiderMonkey now reports more branches: for loop conditions and for
default cases in switch statements.

The gjs_parse_call_args() tests depend on the text of certain error
messages, which were changed in SpiderMonkey 78.

See: GNOME/gjs#329