arg-cache: Use gjs_arg_set_from_js_value to set arguments from JS
We already have a function to set a GArgument from the JS value, taking care
of number conversions and out of bound errors, that is used in arg in the
same way, so let's use this function to be consistent, ensuring that we can
save each type in a container type big enough and that, in case we go out of
bounds we can handle this correctly.

Only use a simple function override, so that we can handle the out of
bound errors the same way we used to do.