coverage: Treat coverage prefix as absolute path
Instead of comparing the coverage prefixes and filenames as strings,
make sure they are treated as absolute paths, if necessary resolved
against the current working directory.

This is a change in the behaviour of the coverage-prefix option, but I
believe it's overall beneficial. The only case that this change would
break, is using an absolute path as the coverage prefix, and a relative
path as the filename, or vice versa, in which case no coverage would
have been output anyway.

Closes: #322