meson: Fix Windows builds against mozjs-78 and later
Since Spidermonkey 78, the headers shipped with it do not work on Visual Studio
builds without major and intrusive changes, so make the Meson build files work
for using CLang's `clang-cl` compiler, which is CLang that tries hard to
emulate the MSVC compiler and is the compiler that is being used to build
SpiderMonkey on Windows nowadays.  As a result, we use Meson's
`cxx.get_argument_syntax() == 'msvc'` rather than `cxx.get_id() === 'msvc'`
(and so for the c compiler), so that we cover the case where we are using
`clang-cl`.

This also means that an error message will be displayed when one tries to use
the Visual Studio compiler when building GJS, and that we can start sharing
some of the compiler flags that is also being used for GCC (and other CLang)
builds.  This also means that we can clean up the build files a little bit.

However, this will still require and use Microsoft's Windows SDK's headers and
libraries, so there are some Visual Studio-specific compiler flags that needs
to be kept, and the Cairo dependencies continue to be searched manually in
cases where the pkg-config files for them cannot be found.

Note that if the other dependencies of GJS were built with Visual Studio, they
do not need to be recompiled with `clang-cl` as long as they link to the same
C Runtime library DLL.